Adjusting tests.
