+++ graylog2-server/src/test/java/org/graylog2/indexer/messages/MessagesIT.java
@@ -34,7 +34,6 @@ import org.graylog2.indexer.retention.strategies.DeletionRetentionStrategyConfig
 import org.graylog2.indexer.rotation.strategies.MessageCountRotationStrategy;
 import org.graylog2.indexer.rotation.strategies.MessageCountRotationStrategyConfig;
 import org.graylog2.plugin.Message;
-import org.graylog2.system.processing.InMemoryProcessingStatusRecorder;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.junit.After;
@@ -51,7 +50,7 @@ import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class MessagesIT extends ElasticsearchBaseTest {
+public abstract class MessagesIT extends ElasticsearchBaseTest {
     private static final String INDEX_NAME = "messages_it_deflector";
 
     private Messages messages;
@@ -77,13 +76,16 @@ public class MessagesIT extends ElasticsearchBaseTest {
             .build();
     private static final IndexSet indexSet = new TestIndexSet(indexSetConfig);
 
+    protected abstract MessagesAdapter createMessagesAdapter(MetricRegistry metricRegistry);
+
     @Before
     public void setUp() throws Exception {
         client().deleteIndices(INDEX_NAME);
         client().createIndex(INDEX_NAME);
         client().waitForGreenStatus(INDEX_NAME);
         count = new Count.Builder().addIndex(INDEX_NAME).build();
-        messages = new Messages(new MetricRegistry(), jestClient(), new InMemoryProcessingStatusRecorder(), true);
+        final MetricRegistry metricRegistry = new MetricRegistry();
+        messages = new Messages(metricRegistry, createMessagesAdapter(metricRegistry));
     }
 
     @After
